Use AC_PROG_CC_C99 to discover C99-compliant compiler.
Instead of just checking if the compiler supports -std=gnu99 option,
make use of the autoconf macro for discovering C99. This way other
non-GCC compiler could be used in C99 mode too.

If the compiler does not support gnu99 it falls back to c99, giving
more chances that it would work than leaving it to default.